+++ actionmailer/lib/action_mailer/gem_version.rb
@@ -10,7 +10,7 @@ module ActionMailer
     MAJOR = 7
     MINOR = 0
     TINY  = 0
-    PRE   = "rc3"
+    PRE   = nil
 
     STRING = [MAJOR, MINOR, TINY, PRE].compact.join(".")
   end